refactor: models subpackage for core and internal models
