+++ docs/source/aind_data_transfer_service.models.rst
@@ -1,0 +1,29 @@
+aind\_data\_transfer\_service.models package
+============================================
+
+Submodules
+----------
+
+aind\_data\_transfer\_service.models.core module
+------------------------------------------------
+
+.. automodule:: aind_data_transfer_service.models.core
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+aind\_data\_transfer\_service.models.internal module
+----------------------------------------------------
+
+.. automodule:: aind_data_transfer_service.models.internal
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: aind_data_transfer_service.models
+   :members:
+   :undoc-members:
+   :show-inheritance:
