+++ src/aind_data_transfer_service/models/__init__.py
@@ -1,0 +1,1 @@
+"""Package for models"""
