+++ src/aind_data_transfer_service/models/core.py
